+++ .travis.yml
@@ -2,11 +2,19 @@ sudo: false
 
 language: rust
 
-rust:
-  - stable
-  - beta
-  - nightly
-  - 1.20.0
-
-script:
-  - cargo test
+matrix:
+  include:
+    - rust: 1.15.0
+      env: RUSTFLAGS="--cfg no_attributes --cfg no_raw_identifiers"
+      script: cargo test --test test
+    - rust: 1.20.0
+      env: RUSTFLAGS="--cfg no_raw_identifiers"
+      script: cargo test
+    - rust: stable
+      env: RUSTFLAGS="--cfg no_raw_identifiers"
+      script: cargo test
+    - rust: beta
+      env: RUSTFLAGS="--cfg no_raw_identifiers"
+      script: cargo test
+    - rust: nightly
+      script: cargo test
