+++ tests/conditional/test_raw_identifiers.rs
@@ -1,0 +1,12 @@
+#[test]
+fn test_raw_identifier() {
+    mashup! {
+        m["x"] = F r#move;
+    }
+
+    m! {
+        struct "x";
+    }
+
+    let _ = Fmove;
+}
